nop worker misc refactor
